Install the gruvbox colorscheme and use it in Neovim and GVim
gruvbox is updated more often than Inkpot and works better in modern Vim - however, the 'termguicolors' option doesn't quite work correctly in Vim and won't give you a transparent terminal background, so we still use Inkpot in that case.

In Neovim, 'termguicolors' works perfectly - and in GVim it's moot since you're not using a terminal anyway. ;)