Support the XDG_STATE_HOME directory
It's a better place to store most of the stuff Vim wants to store lying
around on the filesystem than XDG_CACHE_HOME, it turns out? I'm still
keeping installed packages primarily in XDG_CACHE_HOME, since if that
directory is blown away then the same packages just get reinstalled
automatically and nothing particularly important is lost, but the other
stuff doesn't really belong in there.

Undo files, viminfo, and so on are definitely a good example of "state
data that should be persisted between (application) restarts, but that
is not important or portable enough to the user that it should be stored
in XDG_DATA_HOME", so XDG_STATE_HOME is a fine place for them.